doc: 更新模板示例
